fix: wrong type of UserRole

+++ src/main/java/com/chaffee/entity/dto/LoginDTO.java
@@ -18,7 +18,7 @@ public class LoginDTO implements Serializable {
   
   private String userPassword;
   
-  private Integer userRole;
+  private Long userRole;
   
   private String avatar;
 }
